plotted semi-definite boundary
