--- hw1-geomviz/symmetric-positive-definite/symmetric-positive-definite.py
@@ -1,20 +1,0 @@
-def plot():
-    pass
-
-def plot_grid():
-    pass
-
-def plot_rendering():
-    pass
-
-def plot_tangent_space():
-    pass
-
-def scatter():
-    pass
-
-def plot_geodesic():
-    pass
-
-def plot_vector_field():
-    pass
