+++ hw1-geomviz/symmetric_positive_definite/README.md
@@ -1,0 +1,2 @@
+source activate myenv
+python -m ipykernel install --user --name myenv --display-name "Python (myenv)"
